Shrink the diff by undoing a gratuitous function move.
