+++ protocol/bc/asset_test.go
@@ -1,7 +1,6 @@
 package bc
 
 import (
-	"encoding/hex"
 	"testing"
 
 	"golang.org/x/crypto/sha3"
@@ -35,15 +34,4 @@ func BenchmarkComputeAssetID(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		assetIDSink = ComputeAssetID(issuanceScript, initialBlockHash, 1, EmptyStringHash)
 	}
-}
-
-func mustDecodeHash(hash string) (h [32]byte) {
-	if len(hash) != hex.EncodedLen(len(h)) {
-		panic("wrong length hash")
-	}
-	_, err := hex.Decode(h[:], []byte(hash))
-	if err != nil {
-		panic(err)
-	}
-	return h
 }
