+++ protocol/bc/bc_test.go
@@ -2,6 +2,7 @@ package bc
 
 import (
 	"bytes"
+	"encoding/hex"
 	"io"
 	"testing"
 )
@@ -13,4 +14,15 @@ func serialize(t *testing.T, wt io.WriterTo) []byte {
 		t.Fatal(err)
 	}
 	return b.Bytes()
+}
+
+func mustDecodeHash(hash string) (h [32]byte) {
+	if len(hash) != hex.EncodedLen(len(h)) {
+		panic("wrong length hash")
+	}
+	_, err := hex.Decode(h[:], []byte(hash))
+	if err != nil {
+		panic(err)
+	}
+	return h
 }
